[level 0] Title: 두 수의 나눗셈, Time: 0.02 ms, Memory: 82.2 MB -BaekjoonHub

+++ 프로그래머스/lv0/120806. 두 수의 나눗셈/두 수의 나눗셈.java
@@ -1,7 +1,8 @@
 class Solution {
     public int solution(int num1, int num2) {
         int answer = 0;
-        answer =(int)((num1/(float)num2)*1000);
+        answer= (int)(((double)num1/num2)*1000);
+        
         return answer;
     }
 }
